Update reduxReactRouter to accept history object
When using `history` basename function it returns an object. Before this
it would only accept a function that returns an object. This update
gives users the ability to use other helpers provided by history.

Also update to the tests to account for this new behavior. Most of the
tests were kept in this describe block to ensure all other routes stayed
the same and that history properly uses the basename for determining root.